requirements.txt
llama-cpp-python==0.2.63
pdfplumber==0.11.4
pypdfium2==4.24.0
pillow==10.3.0
pytesseract==0.3.10
PyYAML==6.0.1
numpy==1.26.4
